refactor: clean up/document patcher code
