+++ packages/vexide-startup/src/patcher/varint_encoding.rs
